delete
delete from shopping card

+++ acesso_movel/snakRestaurant/app/src/main/java/amsi/dei/estg/ipleiria/snakrestaurant/models/Shopping_card.java
@@ -3,7 +3,7 @@ package amsi.dei.estg.ipleiria.snakrestaurant.models;
 public class Shopping_card {
 
     private int id_shopping;
-    private int id_product_shopping;
+    private long id_product_shopping;
     private String name_shopping;
     private int price_shopping;
     public int id_category_shopping;
@@ -11,7 +11,7 @@ public class Shopping_card {
 
     //ID_PRODUCT_SHOPPING, NAME_SHOPPING, PRICE_SHOPPING, ID_CATEGORY_SHOPPING, QUANTIDADE
 
-    public Shopping_card(int id_shopping,int id_product_shopping, String name_shopping, int price_shopping, int id_category_shopping, int quantidade_shopping) {
+    public Shopping_card(int id_shopping, long id_product_shopping, String name_shopping, int price_shopping, int id_category_shopping, int quantidade_shopping) {
         this.id_shopping = id_shopping;
         this.id_product_shopping = id_product_shopping;
         this.name_shopping = name_shopping;
@@ -19,6 +19,7 @@ public class Shopping_card {
         this.id_category_shopping = id_category_shopping;
         this.quantidade_shopping = quantidade_shopping;
     }
+
 
     public int getId_shopping() {
         return id_shopping;
@@ -28,11 +29,11 @@ public class Shopping_card {
         this.id_shopping = id_shopping;
     }
 
-    public int getId_product_shopping() {
+    public long getId_product_shopping() {
         return id_product_shopping;
     }
 
-    public void setId_product_shopping(int id_product_shopping) {
+    public void setId_product_shopping(long id_product_shopping) {
         this.id_product_shopping = id_product_shopping;
     }
